Fix event handlers
Update the event handlers to be registered in one class and fix the two block break events incorrectly registered
